Moved responses structs to namespace files (#19)
