--- responses/conduit_connect.go
@@ -1,7 +1,0 @@
-package responses
-
-// ConduitConnectResponse represents the response from calling conduit.connect.
-type ConduitConnectResponse struct {
-	SessionKey   string `json:"sessionKey"`
-	ConnectionID int64  `json:"connectionID"`
-}
