--- responses/differential_getcommitmessage.go
@@ -1,5 +1,0 @@
-package responses
-
-// DifferentialGetCommitMessageResponse is the response of calling
-// differential.getcommitmessage.
-type DifferentialGetCommitMessageResponse string
