--- responses/differential_getcommitpaths.go
@@ -1,5 +1,0 @@
-package responses
-
-// DifferentialGetCommitPathsResponse is the response of calling
-// differential.getcommitpaths.
-type DifferentialGetCommitPathsResponse []string
