--- responses/differential_query.go
@@ -1,6 +1,0 @@
-package responses
-
-import "github.com/uber/gonduit/entities"
-
-// DifferentialQueryResponse is the response of calling differential.query.
-type DifferentialQueryResponse []*entities.DifferentialRevision
