--- responses/differential_querydiffs.go
@@ -1,6 +1,0 @@
-package responses
-
-import "github.com/uber/gonduit/entities"
-
-// DifferentialQueryDiffsResponse is the response of calling differential.querydiffs.
-type DifferentialQueryDiffsResponse []*entities.DifferentialDiff
