+++ responses/diffusion.go
