--- responses/maniphest_gettasktransactions.go
@@ -1,6 +1,0 @@
-package responses
-
-import "github.com/uber/gonduit/entities"
-
-// ManiphestGetTaskTransactionsResponse is the response of calling maniphest.query.
-type ManiphestGetTaskTransactionsResponse map[string][]*entities.ManiphestTaskTranscation
