--- responses/maniphest_query.go
@@ -1,15 +1,0 @@
-package responses
-
-import "github.com/uber/gonduit/entities"
-
-// ManiphestQueryResponse is the response of calling maniphest.query.
-type ManiphestQueryResponse map[string]*entities.ManiphestTask
-
-// Get gets the task with the speicfied numeric ID.
-func (res ManiphestQueryResponse) Get(key string) *entities.ManiphestTask {
-	if _, ok := res[key]; ok {
-		return res[key]
-	}
-
-	return nil
-}
