--- responses/paste_create.go
@@ -1,6 +1,0 @@
-package responses
-
-import "github.com/uber/gonduit/entities"
-
-// PasteCreateResponse represents the result of calling paste.create.
-type PasteCreateResponse *entities.PasteItem
