--- responses/phid_query.go
@@ -1,6 +1,0 @@
-package responses
-
-import "github.com/uber/gonduit/entities"
-
-// PHIDQueryResponse is the result of phid.query operations.
-type PHIDQueryResponse map[string]*entities.PHIDResult
